refactor: convert tempo binary into a library
Move the bulk of `main.rs` into a new `lib.rs`, exposing the launch
logic as `pub fn tempo_main()`. CLI type definitions (`TempoCli`,
`TempoArgs`, `TempoRpcModuleValidator`, `PyroscopeArgs`, the custom RPC
module list and `NodeCommandExt`) move into a new `cli` module. `main.rs`
is now a thin entrypoint that keeps the global allocator and jemalloc
profiling configuration and calls `tempo::tempo_main()`.

This is a pure restructuring with no behavior change, enabling the launch
logic and CLI types to be reused as a library.
